Implement first version of parser
Includes an example `parselog` which prints all of the actions a Parser
implementation is given the opportunity to handle. One way to test this
is to pipe vim into it:

    vim | target/release/examples/parselog

And type `:q` to quit. Vim won't show up, but it still accepts input.

This version of the parser doesn't handle UTF-8. It's implemented as
described by http://vt100.net/emu/dec_ansi_parser which did not include
UTF-8 support.

Next steps are adding UTF-8 support.