Fix off-by-one error when parsing CSI escapes
The CSI escape sequences are limited to 16 parameters using the
`MAX_PARAMS` constant.

To make sure no more params are parsed, the CSI parser stops when the
parameter index reaches `MAX_PARAMS`, however this means that it stops
at index 16, which means that the number of params would be 17 since
arrays start at 0.

To fix this off-by-one error the CSI parser now stops when the index
reaches `MAX_PARAMS - 1`, which means it will stop when the index
reaches the last element in an array with `MAX_PARAMS` number of
elements.

This fixes jwilm/alacritty#1505.