Make all methods of the trait `Perform` optional
I've noticed while playing around with vte in a personal project, that I
didn't need all of the methods of the `Perform` trait. In Alacritty we
also don't react to everything and other crates like
`strip-ansi-escapes` basically doesn't respond to anything.

Of course it's always easy to just copy/paste the entire trait and move
on, but I think it's probably worth making the life of downstream easier
by not enforcing this.