Allow cloning `utf8parser::Parser`
I am using this to validate output sent to `std::io::Write::write` but
the entire validated output might not be written.  To workaround this, I
need to snapshot my validation state, attempt the `write`, and then
replay my validation up-to what was written.  This requires the data to
be `clone`able to do so but `Parser` isn't.

Much like `std::ops::Range`, we likely don't want this to be `Copy` as
that makes it too easy to get mixed up on what state you are using but
`Clone` should be explicit enough to be safe.