ansi: remove the `NullTimeProvider`
Instead of providing a "broken" sync handler for no_std by default, we
now force the user to specify the sync handler. This would make sure
that the users know what is going on without relying on "broken" sync
handler implementation. If they want it, its pretty simple to make a
stub impl anyways so yeah. :)

Signed-off-by: Andy-Python-Programmer <andypythonappdeveloper@gmail.com>