fix ub unpack routine
This patch fixes the undefined behavior in the
unpack routine resulting from the fact that
enum representation is undefined. Previously,
the code assumed that the enum values were represented
in the sensible way, which technically it had no
way to be sure of. With the repr tags, it is ok to
cast directly to the enum values.