Rewrite table based state change to `match` based
The table based state change was too complex to make guesses why it's
getting slow and too fragile, as in modifying the amount of
states/actions were slowing down, even though, they were not used.

Rewrite the state + action change exactly how it's in [1] with respect
to our modifications/C1, etc. The new implementation is generally faster
than the previous one and is easier for compiler to reason about and
generate more efficient structures.

Also, the structure got way simpler to follow, since it matches the
spec pretty much exactly.

[1] - https://vt100.net/emu/dec_ansi_parser