Replace `no_std` with `std` feature
The `no_std` feature has been broken since version `0.14`, since it did
not disable the `std` feature of `memchr`.

To fall more in line with the general ecosystem standards, this crate
changes the `no_std` feature into an `std` feature. Since `vte` should
be significantly faster with this feature, it has also been enabled by
default (previously `no_std` was default).

This is a breaking change, even if we ignore `no_std` consumers which
aren't able to compile 0.14, since intermediate dependencies relying on
the `no_std` feature will have to remove it from the manifest.